+++ api/core/src/modules/permissions/permissions.module.ts
@@ -1,8 +1,10 @@
-import { Module } from "@nestjs/common";
+import { Global, Module } from "@nestjs/common";
+import { PermissionGuard } from "../../common/guards/permission.guard";
 import { PermissionsService } from "./permissions.service";
 
+@Global()
 @Module({
-  providers: [PermissionsService],
-  exports: [PermissionsService]
+  providers: [PermissionGuard, PermissionsService],
+  exports: [PermissionGuard, PermissionsService]
 })
 export class PermissionsModule {}